\name{envData}
\alias{envData}
\docType{data}

\title{The variables in the environment variable \code{envData} of the system}
\description{
  The variables in the environment variable \code{envData} of the system.
}

\format{An environment variable}
\author{Xinrui Shi, Chunquan Li and Xia Li}
\keyword{file}